Tabix indexing of empty filtered FreeBayes variant files

+++ bcbio/variation/vfilter.py
@@ -135,7 +135,10 @@ def _freebayes_hard(in_file, data):
     if not vcfutils.vcf_has_variants(in_file):
         base, ext = utils.splitext_plus(in_file)
         out_file = "{base}-filter{ext}".format(**locals())
-        shutil.copy(in_file, out_file)
+        if not utils.file_exists(out_file):
+            shutil.copy(in_file, out_file)
+        if out_file.endswith(".vcf.gz"):
+            out_file = vcfutils.bgzip_and_index(out_file, data["config"])
         return out_file
     
     stats = _calc_vcf_stats(in_file)
